@@ -17,34 +17,35 @@
     "url": "https://github.com/feedm3/unhypem/issues"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.0",
-    "babel-loader": "^6.2.7",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
     "babel-plugin-lodash": "^3.1.4",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-register": "^6.18.0",
-    "css-loader": "^0.26.2",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.24.1",
+    "css-loader": "^0.28.4",
     "dotenv": "^4.0.0",
-    "eslint": "^3.16.1",
-    "eslint-plugin-react": "^6.2.0",
-    "eslint-plugin-standard": "^2.0.0",
-    "extract-text-webpack-plugin": "^2.0.0",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-standard": "^3.0.1",
+    "extract-text-webpack-plugin": "^2.1.0",
     "flux": "^3.1.2",
     "lodash": "^4.13.1",
-    "moment": "^2.15.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.1.0",
-    "react-router": "^3.0.2",
-    "react-tether": "^0.5.6",
-    "request": "^2.76.0",
+    "moment": "^2.18.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-router": "^3.0.5",
+    "react-tether": "^0.5.7",
+    "request": "^2.81.0",
     "soundmanager2": "^2.97.20150601-a",
-    "style-loader": "^0.13.2",
-    "superagent": "^3.5.0",
-    "svg-sprite-loader": "^0.3.0",
+    "style-loader": "^0.18.2",
+    "superagent": "^3.5.2",
+    "svg-sprite-loader": "^3.0.5",
     "tether-tooltip": "^1.2.0",
     "url-loader": "^0.5.8",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.1"
+    "webpack": "^2.6.1",
+    "webpack-dev-server": "^2.4.5"
   }
 }
