@@ -5,6 +5,7 @@
 'use strict';
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import TetherComponent from 'react-tether';
 import SvgIcon from '../common/svg-icon';
 import shareIcon from '../../../assets/img/ic_share_black_24px.svg';
@@ -37,7 +38,7 @@ export default class ShareButton extends React.Component {
                         attachment: 'both'
                     }]}>
                     <SvgIcon
-                        id={ shareIcon }
+                        symbol={ shareIcon }
                         title='Share'
                         width='24px'
                         height='24px'
@@ -69,6 +70,6 @@ export default class ShareButton extends React.Component {
     }
 }
 ShareButton.propTypes = {
-    soundcloudUrl: React.PropTypes.string,
-    hypemMediaId: React.PropTypes.string
+    soundcloudUrl: PropTypes.string,
+    hypemMediaId: PropTypes.string
 };
