@@ -13,7 +13,7 @@ import rewindIcon from '../../../assets/img/ic_skip_previous_black_24px.svg';
 export default () => {
     return (
         <SvgIcon
-            id={rewindIcon}
+            symbol={ rewindIcon }
             title="Rewind"
             width="36px"
             height="36px"
